+++ box-ui/src/components/ts/JitsiFrame.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState, useEffect, FunctionComponent } from 'react';
+import React, { useEffect, FunctionComponent } from 'react';
 import '../css/JitsiComponent.css';
 import { JitsiFrameProps } from '../../utils/Props';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -51,10 +51,6 @@ const JitsiFrame: FunctionComponent<JitsiFrameProps> = (props: JitsiFrameProps) 
                     api.dispose();
                 });
                 props.configure?.(api);
-                api.addListener('raiseHandUpdated', (res) => {
-                    const timeRaised = res.handRaised;
-                    updateCounter(timeRaised);
-                });
             })
             .catch(() => {
                 alert('Error loading Jitsi Meet API');
@@ -67,54 +63,9 @@ const JitsiFrame: FunctionComponent<JitsiFrameProps> = (props: JitsiFrameProps) 
         };
     }, [props.information, props.options, props.configure, props.onError]);
 
-    const [raised, setRaised] = useState(false);
-    const [counter, setCounter] = useState(0);
-    const TimeOutRef = useRef<NodeJS.Timeout>();
-
-    function lowerHand() {
-        setRaised(false);
-        setCounter(0);
-    }
-
-    function updateCounter(timeRaised: number) {
-        if (timeRaised > 0) {
-            setCounter((counter) => counter + 1);
-        }
-        if (timeRaised == 0) {
-            setCounter((counter) => Math.max(0, counter - 1));
-        }
-    }
-
-    function switchHand() {
-        let stop = false;
-        if (counter > 0) {
-            stop = true;
-        }
-        if (stop) {
-            setRaised(true);
-        } else {
-            setRaised(false);
-        }
-    }
-    useEffect(() => {
-        const id = setTimeout(lowerHand, 10000);
-        if (TimeOutRef.current) {
-            clearTimeout(TimeOutRef.current);
-        }
-        TimeOutRef.current = id;
-        switchHand();
-    }, [counter]);
-
     return (
         <div style={{ height: '100%' }}>
             <div id='jitsi-container' className='jitsiContainer' />
-            <div
-                className='overlay'
-                style={{
-                    backgroundColor: 'yellow',
-                    opacity: raised ? '0.5' : '0',
-                }}
-            />
         </div>
     );
 };
